feat(csharp/src/Drivers/Apache): make Apace driver tests inheritable
