+++ csharp/test/Drivers/Apache/Spark/StringValueTests.cs
@@ -15,127 +15,52 @@
 * limitations under the License.
 */
 
-using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
-using Apache.Arrow.Adbc.Drivers.Apache.Hive2;
 using Apache.Arrow.Adbc.Drivers.Apache.Spark;
 using Xunit;
 using Xunit.Abstractions;
 
 namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
 {
-    // TODO: When supported, use prepared statements instead of SQL string literals
-    //      Which will better test how the driver handles values sent/received
-
-    /// <summary>
-    /// Validates that specific string and character values can be inserted, retrieved and targeted correctly
-    /// </summary>
-    public class StringValueTests : TestBase<SparkTestConfiguration, SparkTestEnvironment>
+    public class StringValueTests(ITestOutputHelper output)
+        : Common.StringValueTests<SparkTestConfiguration, SparkTestEnvironment>(output, new SparkTestEnvironment.Factory())
     {
-        public StringValueTests(ITestOutputHelper output) : base(output, new SparkTestEnvironment.Factory()) { }
-
-        public static IEnumerable<object[]> ByteArrayData(int size)
-        {
-            var rnd = new Random();
-            byte[] bytes = new byte[size];
-            rnd.NextBytes(bytes);
-            yield return new object[] { bytes };
-        }
-
-        /// <summary>
-        /// Validates if driver can send and receive specific String values correctly.
-        /// </summary>
         [SkippableTheory]
-        [InlineData(null)]
-        [InlineData("")]
-        [InlineData("你好")]
         [InlineData("String contains formatting characters tab\t, newline\n, carriage return\r.", SparkServerType.Databricks)]
-        [InlineData(" Leading and trailing spaces ")]
-        internal async Task TestStringData(string? value, SparkServerType? serverType = default)
+        internal async Task TestStringDataDatabricks(string? value, SparkServerType serverType)
         {
-            Skip.If(serverType != null && TestEnvironment.ServerType != serverType);
-            string columnName = "STRINGTYPE";
-            using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} {1}", columnName, "STRING"));
-            await ValidateInsertSelectDeleteSingleValueAsync(
-                table.TableName,
-                columnName,
-                value,
-                value != null ? QuoteValue(value) : value);
+            Skip.If(TestEnvironment.ServerType != serverType);
+            await TestStringData(value);
         }
 
-        /// <summary>
-        /// Validates if driver can send and receive specific VARCHAR values correctly.
-        /// </summary>
         [SkippableTheory]
-        [InlineData(null)]
-        [InlineData("")]
-        [InlineData("你好")]
         [InlineData("String contains formatting characters tab\t, newline\n, carriage return\r.", SparkServerType.Databricks)]
-        [InlineData(" Leading and trailing spaces ")]
-        internal async Task TestVarcharData(string? value, SparkServerType? serverType = default)
+        internal async Task TestVarcharDataDatabricks(string? value, SparkServerType serverType)
         {
-            Skip.If(serverType != null && TestEnvironment.ServerType != serverType);
-            string columnName = "VARCHARTYPE";
-            using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} {1}", columnName, "VARCHAR(100)"));
-            await ValidateInsertSelectDeleteSingleValueAsync(
-                table.TableName,
-                columnName,
-                value,
-                value != null ? QuoteValue(value) : value);
+            Skip.If(TestEnvironment.ServerType != serverType);
+            await TestVarcharData(value);
         }
 
-        private bool IsBelowMinimumVersion(string? minVersion) => minVersion != null && VendorVersionAsVersion < Version.Parse(minVersion);
-
-        /// <summary>
-        /// Validates if driver can send and receive specific VARCHAR values correctly.
-        /// </summary>
         [SkippableTheory]
-        [InlineData(null)]
-        [InlineData("")]
-        [InlineData("你好")]
         [InlineData("String contains formatting characters tab\t, newline\n, carriage return\r.", SparkServerType.Databricks)]
-        [InlineData(" Leading and trailing spaces ")]
-        internal async Task TestCharData(string? value, SparkServerType? serverType = default)
+        internal async Task TestCharDataDatabricks(string? value, SparkServerType serverType)
         {
-            Skip.If(serverType != null && TestEnvironment.ServerType != serverType);
-            string columnName = "CHARTYPE";
-            int fieldLength = 100;
-            using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} {1}", columnName, $"CHAR({fieldLength})"));
-
-            string? formattedValue = value != null ? QuoteValue(value.PadRight(fieldLength)) : value;
-            string? paddedValue = value != null ? value.PadRight(fieldLength) : value;
-
-            await InsertSingleValueAsync(table.TableName, columnName, formattedValue);
-            await SelectAndValidateValuesAsync(table.TableName, columnName, paddedValue, 1, formattedValue);
-            string whereClause = GetWhereClause(columnName, formattedValue ?? paddedValue);
-            if (SupportsDelete) await DeleteFromTableAsync(table.TableName, whereClause, 1);
+            Skip.If(TestEnvironment.ServerType != serverType);
+            await TestCharData(value);
         }
 
-        /// <summary>
-        /// Validates if driver fails to insert invalid length of VARCHAR value.
-        /// </summary>
+        protected override async Task TestVarcharExceptionData(string value, string[] expectedTexts, string? expectedSqlState)
+        {
+            Skip.If(TestEnvironment.ServerType == SparkServerType.Databricks);
+            await base.TestVarcharExceptionData(value, expectedTexts, expectedSqlState);
+        }
+
         [SkippableTheory]
-        [InlineData("String whose length is too long for VARCHAR(10).")]
-        public async Task TestVarcharExceptionData(string value)
+        [InlineData("String whose length is too long for VARCHAR(10).", new string[] { "DELTA_EXCEED_CHAR_VARCHAR_LIMIT", "DeltaInvariantViolationException" }, "22001")]
+        public async Task TestVarcharExceptionDataDatabricks(string value, string[] expectedTexts, string? expectedSqlState)
         {
-            string columnName = "VARCHARTYPE";
-            using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} {1}", columnName, "VARCHAR(10)"));
-            AdbcException exception = await Assert.ThrowsAsync<HiveServer2Exception>(async () => await ValidateInsertSelectDeleteSingleValueAsync(
-                table.TableName,
-                columnName,
-                value,
-                value != null ? QuoteValue(value) : value));
-
-            bool serverTypeDatabricks = TestEnvironment.ServerType == SparkServerType.Databricks;
-            string[] expectedTexts = serverTypeDatabricks
-                ? ["DELTA_EXCEED_CHAR_VARCHAR_LIMIT", "DeltaInvariantViolationException"]
-                : ["Exceeds", "length limitation: 10"];
-            AssertContainsAll(expectedTexts, exception.Message);
-
-            string? expectedSqlState = serverTypeDatabricks ? "22001" : null;
-            Assert.Equal(expectedSqlState, exception.SqlState);
+            Skip.IfNot(TestEnvironment.ServerType == SparkServerType.Databricks, $"Server type: {TestEnvironment.ServerType}");
+            await base.TestVarcharExceptionData(value, expectedTexts, expectedSqlState);
         }
-
     }
 }
